+++ src/FBCModelTests.jl
@@ -1,9 +1,3 @@
-#=
-This package contains a collection of tests based on Memote. See Lieven, C., Beber,
-M.E., Olivier, B.G. et al. MEMOTE for standardized genome-scale metabolic model
-testing. Nat Biotechnol 38, 272–276 (2020).
-https://doi.org/10.1038/s41587-020-0446-y for details.
-=#
 """
 # FBCModelTests v$(FBCModelTests.FBCMT_VERSION)
 
@@ -19,9 +13,6 @@ include("Utils.jl")
 import .Utils
 
 # include(joinpath("frog", "FROG.jl"))
-
-include(joinpath("memote", "Config.jl"))
-import .Config
 
 include(joinpath("memote", "Memote.jl"))
 import .Memote
