+++ src/memote/checks/Annotation.jl
@@ -1,5 +1,5 @@
 """
-$(TYPEDEF)
+module Annotation
 
 This module contains tests used to check the coverage and conformance of
 reaction, gene, and metabolite annotations.
